feat: add .gitignore file and create necessary database files and tables
Added a .gitignore file to exclude the "postgres-data" directory from version control. This directory will be used to store PostgreSQL data files.

Created the following database files and tables:
- db.go: Added the openDB function to establish a connection to the PostgreSQL database using the provided DSN (Data Source Name).
- main.go: Added the DSN flag to read the PostgreSQL connection string from the command line. Also added the connectToDB function to establish a connection to the database and store it in the application struct.
- docker-compose.yml: Added a Docker Compose file to define the PostgreSQL service and its configuration.
- create_tables.sql: Added the SQL script to create the necessary tables in the PostgreSQL database.

These changes were made to set up the database infrastructure and ensure that the application can connect to the PostgreSQL database.